feat: Enable and configure local build cache
This commit enables the local build cache in `settings.gradle.kts` and configures it to be stored in the `gradle-cache` directory at the root of the project.

The `build.sh` script has been updated to automatically add and commit the `gradle-cache` directory if it exists.

The `.gitignore` file has been modified to stop ignoring the `gradle-cache` directory.